refactor: fold fee-policy allowlist actions into NetworkAccountConfig note
Now that FeePolicyManager is part of AuthNetworkAccount (not a standalone
component), manage the allowed fee policy roots through the network account's
own config note rather than a separate note.

- Remove the standalone FeePolicyManagerConfig note (masm + Rust + registrations).
- Extend NetworkAccountConfig with two selectors, AddAllowedFeePolicy (4) and
  RemoveAllowedFeePolicy (5), dispatching to fee_manager::add_allowed_fee_policy /
  remove_allowed_fee_policy. The note script's root loader is generalized from
  load_script_root to load_root since it now loads note-script, tx-script, or
  fee-policy roots.
