--- crates/miden-standards/asm/standards/notes/fee_policy_manager_config.masm
@@ -1,157 +1,0 @@
-use miden::protocol::active_note
-use miden::standards::fees::fee_manager
-
-# CONSTANTS
-# =================================================================================================
-
-# Memory base at which the note storage is written by `active_note::get_storage`.
-const STORAGE_PTR = 0
-
-# Memory addresses of the FeePolicyManagerConfig note storage. The fee policy root word occupies the
-# first four items with the selector after it.
-const POLICY_ROOT_PTR = STORAGE_PTR
-const SELECTOR_PTR = STORAGE_PTR + 4
-
-# Action selectors, stored in the storage item after the policy root. Keep in sync with
-# `FeePolicyManagerConfig` on the Rust side.
-const SELECTOR_ADD_ALLOWED_FEE_POLICY = 0
-const SELECTOR_REMOVE_ALLOWED_FEE_POLICY = 1
-
-# Expected note storage item count: a selector plus the policy root word. Identical for every
-# action.
-const NUM_STORAGE_ITEMS = 5
-
-# ERRORS
-# =================================================================================================
-
-const ERR_FEE_POLICY_MANAGER_CONFIG_UNKNOWN_SELECTOR = "fee policy manager config note selector does not match a known action"
-
-const ERR_FEE_POLICY_MANAGER_CONFIG_UNEXPECTED_NUMBER_OF_STORAGE_ITEMS = "storage item count of the fee policy manager config note does not match the expected number"
-
-# NOTE SCRIPT
-# =================================================================================================
-
-#! FeePolicyManagerConfig script: mutates the allowed fee policy roots map, dispatched on a selector
-#! stored in the first note storage item.
-#!
-#! The selected action and its fee policy root are read from the note storage. They are fixed at
-#! note creation, so authorization is bound to the note sender. The storage is laid out as
-#! `[POLICY_ROOT, selector]`:
-#! - `[POLICY_ROOT, 0]` AddAllowedFeePolicy.
-#! - `[POLICY_ROOT, 1]` RemoveAllowedFeePolicy.
-#!
-#! Authorization is enforced by the called fee policy procedures via the account-wide `Authority`
-#! component, which resolves the note sender against the configured owner or role.
-#!
-#! Requires that the account exposes the `miden::standards::fees::fee_manager` procedures (as the
-#! `AuthNetworkAccount` component does), namely:
-#! - `miden::standards::fees::fee_manager::add_allowed_fee_policy`
-#! - `miden::standards::fees::fee_manager::remove_allowed_fee_policy`
-#!
-#! Inputs:  [ARGS, pad(12)]
-#! Outputs: [pad(16)]
-#!
-#! Panics if:
-#! - the selector is not a known action.
-#! - the number of storage items does not match the selected action.
-#! - the note sender is not authorized (per the fee policy procedures).
-@note_script
-pub proc main
-    # drop the note script args
-    dropw
-    # => [pad(16)]
-
-    # write the note storage to memory starting at STORAGE_PTR
-    push.STORAGE_PTR exec.active_note::get_storage
-    # => [num_storage_items, pad(16)]
-
-    mem_load.SELECTOR_PTR
-    # => [selector, num_storage_items, pad(16)]
-
-    # dispatch on the selector to the matching action procedure
-    dup eq.SELECTOR_ADD_ALLOWED_FEE_POLICY
-    # => [is_add, selector, num_storage_items, pad(16)]
-
-    if.true
-        drop
-        # => [num_storage_items, pad(16)]
-
-        exec.execute_add_allowed_fee_policy
-    else
-        eq.SELECTOR_REMOVE_ALLOWED_FEE_POLICY
-        assert.err=ERR_FEE_POLICY_MANAGER_CONFIG_UNKNOWN_SELECTOR
-        # => [num_storage_items, pad(16)]
-
-        exec.execute_remove_allowed_fee_policy
-    end
-    # => [pad(16)]
-end
-
-# HELPER PROCEDURES
-# =================================================================================================
-
-#! Asserts the storage item count and loads the stored fee policy root onto the stack.
-#!
-#! Inputs:  [num_storage_items, pad(16)]
-#! Outputs: [POLICY_ROOT, pad(12)]
-#!
-#! Where:
-#! - num_storage_items is the actual number of storage items in the note.
-#! - POLICY_ROOT is the fee policy root read from the note storage.
-#!
-#! Panics if:
-#! - num_storage_items is not equal to the expected number.
-#!
-#! Invocation: exec
-proc load_policy_root
-    eq.NUM_STORAGE_ITEMS
-    assert.err=ERR_FEE_POLICY_MANAGER_CONFIG_UNEXPECTED_NUMBER_OF_STORAGE_ITEMS
-    # => [pad(16)]
-
-    # load the fee policy root
-    mem_loadw_le.POLICY_ROOT_PTR
-    # => [POLICY_ROOT, pad(12)]
-end
-
-#! Executes the AddAllowedFeePolicy action: adds the stored root to the allowed fee policy roots map.
-#!
-#! Inputs:  [num_storage_items, pad(16)]
-#! Outputs: [pad(16)]
-#!
-#! Where:
-#! - num_storage_items is the actual number of storage items in the note.
-#!
-#! Panics if:
-#! - num_storage_items is not equal to the expected number.
-#! - the note sender is not authorized (per `fee_manager::add_allowed_fee_policy`).
-#!
-#! Invocation: exec
-proc execute_add_allowed_fee_policy
-    exec.load_policy_root
-    # => [POLICY_ROOT, pad(12)]
-
-    call.fee_manager::add_allowed_fee_policy
-    # => [pad(16)]
-end
-
-#! Executes the RemoveAllowedFeePolicy action: removes the stored root from the allowed fee policy
-#! roots map.
-#!
-#! Inputs:  [num_storage_items, pad(16)]
-#! Outputs: [pad(16)]
-#!
-#! Where:
-#! - num_storage_items is the actual number of storage items in the note.
-#!
-#! Panics if:
-#! - num_storage_items is not equal to the expected number.
-#! - the note sender is not authorized (per `fee_manager::remove_allowed_fee_policy`).
-#!
-#! Invocation: exec
-proc execute_remove_allowed_fee_policy
-    exec.load_policy_root
-    # => [POLICY_ROOT, pad(12)]
-
-    call.fee_manager::remove_allowed_fee_policy
-    # => [pad(16)]
-end
