+++ crates/miden-standards/src/note/mod.rs
@@ -13,9 +13,6 @@ pub use burn::BurnNote;
 mod faucet_policy_action;
 pub use faucet_policy_action::{FaucetPolicyAction, FaucetPolicyActionNote};
 
-mod fee_policy_manager_config;
-pub use fee_policy_manager_config::{FeePolicyManagerConfig, FeePolicyManagerConfigNote};
-
 mod fee_sponsorship;
 pub use fee_sponsorship::{FeeSponsorshipNote, FeeSponsorshipNoteStorage};
 
@@ -81,7 +78,6 @@ pub enum StandardNote {
     OWNER_ACTION,
     RBAC_ACTION,
     NETWORK_ACCOUNT_CONFIG,
-    FEE_POLICY_MANAGER_CONFIG,
     FEE_SPONSORSHIP,
     TX_FEE,
 }
@@ -132,9 +128,6 @@ impl StandardNote {
         if root == NetworkAccountConfigNote::script_root() {
             return Some(Self::NETWORK_ACCOUNT_CONFIG);
         }
-        if root == FeePolicyManagerConfigNote::script_root() {
-            return Some(Self::FEE_POLICY_MANAGER_CONFIG);
-        }
         if root == FeeSponsorshipNote::script_root() {
             return Some(Self::FEE_SPONSORSHIP);
         }
@@ -162,7 +155,6 @@ impl StandardNote {
             Self::OWNER_ACTION => "OWNER_ACTION",
             Self::RBAC_ACTION => "RBAC_ACTION",
             Self::NETWORK_ACCOUNT_CONFIG => "NETWORK_ACCOUNT_CONFIG",
-            Self::FEE_POLICY_MANAGER_CONFIG => "FEE_POLICY_MANAGER_CONFIG",
             Self::FEE_SPONSORSHIP => "FEE_SPONSORSHIP",
             Self::TX_FEE => "TX_FEE",
         }
@@ -184,7 +176,6 @@ impl StandardNote {
             // RbacAction storage is variable per action; this returns the upper bound.
             Self::RBAC_ACTION => RbacActionNote::MAX_NUM_STORAGE_ITEMS,
             Self::NETWORK_ACCOUNT_CONFIG => NetworkAccountConfigNote::NUM_STORAGE_ITEMS,
-            Self::FEE_POLICY_MANAGER_CONFIG => FeePolicyManagerConfigNote::NUM_STORAGE_ITEMS,
             Self::FEE_SPONSORSHIP => FeeSponsorshipNote::NUM_STORAGE_ITEMS,
             Self::TX_FEE => TxFeeNote::NUM_STORAGE_ITEMS,
         }
@@ -204,7 +195,6 @@ impl StandardNote {
             Self::OWNER_ACTION => OwnerActionNote::script(),
             Self::RBAC_ACTION => RbacActionNote::script(),
             Self::NETWORK_ACCOUNT_CONFIG => NetworkAccountConfigNote::script(),
-            Self::FEE_POLICY_MANAGER_CONFIG => FeePolicyManagerConfigNote::script(),
             Self::FEE_SPONSORSHIP => FeeSponsorshipNote::script(),
             Self::TX_FEE => TxFeeNote::script(),
         }
@@ -224,7 +214,6 @@ impl StandardNote {
             Self::OWNER_ACTION => OwnerActionNote::script_root(),
             Self::RBAC_ACTION => RbacActionNote::script_root(),
             Self::NETWORK_ACCOUNT_CONFIG => NetworkAccountConfigNote::script_root(),
-            Self::FEE_POLICY_MANAGER_CONFIG => FeePolicyManagerConfigNote::script_root(),
             Self::FEE_SPONSORSHIP => FeeSponsorshipNote::script_root(),
             Self::TX_FEE => TxFeeNote::script_root(),
         }
