+++ crates/miden-standards/src/note/network_account_config.rs
@@ -1,6 +1,6 @@
 use alloc::vec::Vec;
 
-use miden_protocol::account::AccountId;
+use miden_protocol::account::{AccountId, AccountProcedureRoot};
 use miden_protocol::assembly::Path;
 use miden_protocol::crypto::rand::FeltRng;
 use miden_protocol::errors::NoteError;
@@ -41,13 +41,13 @@ static NETWORK_ACCOUNT_CONFIG_SCRIPT: LazyLock<NoteScript> = LazyLock::new(|| {
 // NETWORK ACCOUNT CONFIG
 // ================================================================================================
 
-/// An allowlist-mutation action of the
+/// A configuration action of the
 /// [`AuthNetworkAccount`](crate::account::auth::AuthNetworkAccount) component that a
 /// [`NetworkAccountConfigNote`] triggers on the network account that consumes it.
 ///
-/// Each variant adds or removes one script root from the note script or tx script allowlist.
-/// Because the allowlist check reads the transaction's initial state, an update only takes effect
-/// from the account's next transaction.
+/// Each variant adds or removes one root from the note-script allowlist, the tx-script allowlist,
+/// or the allowed fee policy roots map. The allowlist checks read the transaction's initial state,
+/// so an update only takes effect from the account's next transaction.
 ///
 /// The action is encoded into the note's storage (see [`NoteStorage`] conversion below). Because
 /// the storage is fixed at note creation and bound into the note commitment, the authorized party
@@ -63,6 +63,10 @@ pub enum NetworkAccountConfig {
     AddAllowedTxScript { script_root: TransactionScriptRoot },
     /// Removes `script_root` from the tx script allowlist.
     RemoveAllowedTxScript { script_root: TransactionScriptRoot },
+    /// Adds `policy_root` to the allowed fee policy roots map.
+    AddAllowedFeePolicy { policy_root: AccountProcedureRoot },
+    /// Removes `policy_root` from the allowed fee policy roots map.
+    RemoveAllowedFeePolicy { policy_root: AccountProcedureRoot },
 }
 
 impl NetworkAccountConfig {
@@ -75,8 +79,10 @@ impl NetworkAccountConfig {
     const SELECTOR_REMOVE_ALLOWED_NOTE_SCRIPT: u8 = 1;
     const SELECTOR_ADD_ALLOWED_TX_SCRIPT: u8 = 2;
     const SELECTOR_REMOVE_ALLOWED_TX_SCRIPT: u8 = 3;
+    const SELECTOR_ADD_ALLOWED_FEE_POLICY: u8 = 4;
+    const SELECTOR_REMOVE_ALLOWED_FEE_POLICY: u8 = 5;
 
-    /// Returns the selector and the affected script root of this action.
+    /// Returns the selector and the affected root of this action.
     fn parts(self) -> (u8, Word) {
         match self {
             NetworkAccountConfig::AddAllowedNoteScript { script_root } => {
@@ -91,10 +97,16 @@ impl NetworkAccountConfig {
             NetworkAccountConfig::RemoveAllowedTxScript { script_root } => {
                 (Self::SELECTOR_REMOVE_ALLOWED_TX_SCRIPT, script_root.as_word())
             },
+            NetworkAccountConfig::AddAllowedFeePolicy { policy_root } => {
+                (Self::SELECTOR_ADD_ALLOWED_FEE_POLICY, policy_root.as_word())
+            },
+            NetworkAccountConfig::RemoveAllowedFeePolicy { policy_root } => {
+                (Self::SELECTOR_REMOVE_ALLOWED_FEE_POLICY, policy_root.as_word())
+            },
         }
     }
 
-    /// Returns the note storage values encoding this action, laid out as `[SCRIPT_ROOT, selector]`.
+    /// Returns the note storage values encoding this action, laid out as `[ROOT, selector]`.
     fn to_storage_values(self) -> Vec<Felt> {
         let (selector, script_root) = self.parts();
         let mut values = Vec::with_capacity(NetworkAccountConfigNote::NUM_STORAGE_ITEMS);
@@ -114,13 +126,13 @@ impl From<NetworkAccountConfig> for NoteStorage {
 // NETWORK ACCOUNT CONFIG NOTE
 // ================================================================================================
 
-/// A NetworkAccountConfig note: adds or removes a script root from a network account's
-/// note script or tx script allowlist.
+/// A NetworkAccountConfig note: adds or removes a root from a network account's note-script
+/// allowlist, tx-script allowlist, or allowed fee policy roots map.
 ///
 /// A single note script dispatches on a selector in the note's storage to one of the
-/// [`AuthNetworkAccount`](crate::account::auth::AuthNetworkAccount) component's allowlist
-/// procedures. Authorization is enforced by those procedures through the account-wide `Authority`
-/// component against the note sender.
+/// [`AuthNetworkAccount`](crate::account::auth::AuthNetworkAccount) component's allowlist or
+/// fee-policy procedures. Authorization is enforced by those procedures through the account-wide
+/// `Authority` component against the note sender.
 ///
 /// For the consuming network account to accept this note, its own script root must be in the
 /// account's note script allowlist. Every
@@ -292,6 +304,10 @@ mod tests {
         TransactionScriptRoot::from_raw(Word::from([seed, seed + 1, seed + 2, seed + 3]))
     }
 
+    fn policy_root(seed: u32) -> AccountProcedureRoot {
+        AccountProcedureRoot::from_raw(Word::from([seed, seed + 1, seed + 2, seed + 3]))
+    }
+
     /// The builder produces a public, asset-less note tagged for the managed network account.
     #[test]
     fn builder_builds_allowlist_action_note() {
@@ -316,11 +332,12 @@ mod tests {
         assert_eq!(note.assets().num_assets(), 0);
     }
 
-    /// Storage is `[selector, SCRIPT_ROOT]` with the selector matching the action kind.
+    /// Storage is `[ROOT, selector]` with the selector matching the action kind.
     #[test]
     fn storage_layout() {
         let note_root = note_root(10);
         let tx_root = tx_root(20);
+        let policy_root = policy_root(30);
 
         let cases = [
             (
@@ -342,6 +359,16 @@ mod tests {
                 NetworkAccountConfig::RemoveAllowedTxScript { script_root: tx_root },
                 NetworkAccountConfig::SELECTOR_REMOVE_ALLOWED_TX_SCRIPT,
                 tx_root.as_word(),
+            ),
+            (
+                NetworkAccountConfig::AddAllowedFeePolicy { policy_root },
+                NetworkAccountConfig::SELECTOR_ADD_ALLOWED_FEE_POLICY,
+                policy_root.as_word(),
+            ),
+            (
+                NetworkAccountConfig::RemoveAllowedFeePolicy { policy_root },
+                NetworkAccountConfig::SELECTOR_REMOVE_ALLOWED_FEE_POLICY,
+                policy_root.as_word(),
             ),
         ];
